Make it compatible and jdk 11 for 2.13.1
