+++ algebird-core/src/main/scala/com/twitter/algebird/Fold.scala
@@ -17,7 +17,6 @@ package com.twitter.algebird
 
 import java.io.Serializable
 import scala.collection.compat._
-import scala.collection.mutable.Builder
 
 /**
  * Folds are first-class representations of "Traversable.foldLeft." They have the nice property that
@@ -181,7 +180,7 @@ final class FoldState[X, -I, +O] private[algebird] (val add: (X, I) => X, val st
  * Additionally, it is recommended that "end" functions not mutate the accumulator in order to
  * support scans (producing a stream of intermediate outputs by calling "end" at each step).
  */
-object Fold {
+object Fold extends CompatFold {
 
   /**
    * "import Fold.applicative" will bring the Applicative instance into scope. See FoldApplicative.
@@ -245,13 +244,13 @@ object Fold {
       }
     }
 
-  /**
-   * Simple Fold that collects elements into a container.
-   */
-  def container[I, C[_]](implicit cbf: Factory[I, C[I]]): Fold[I, C[I]] =
-    Fold.foldMutable[Builder[I, C[I]], I, C[I]]({ case (b, i) => b += i }, { _ =>
-      cbf.newBuilder
-    }, { _.result })
+  // /**
+  //  * Simple Fold that collects elements into a container.
+  //  */
+  // def container[I, C[_]](implicit cbf: Factory[I, C[I]]): Fold[I, C[I]] =
+  //   Fold.foldMutable[Builder[I, C[I]], I, C[I]]({ case (b, i) => b += i }, { _ =>
+  //     cbf.newBuilder
+  //   }, { _.result })
 
   /**
    * An even simpler Fold that collects into a Seq.  Shorthand for "container[I, Seq];" fewer type
